test: align replay runtime module topology
