+++ src/daemon/handlers/__tests__/session-replay-runtime-plan.test.ts
@@ -9,9 +9,7 @@ import fs from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
 import { runReplayScriptFile } from '../session-replay-runtime.ts';
-import { buildReplayFailureDivergence } from '../session-replay-divergence.ts';
 import { SessionStore } from '../../session-store.ts';
-import type { DaemonRequest, DaemonResponse } from '../../types.ts';
 import { dispatchCommand } from '../../../core/dispatch.ts';
 import { makeIosSession } from '../../../__tests__/test-utils/session-factories.ts';
 import { baseReplayRequest as baseReq, writeReplayFile } from './session-replay-runtime.fixtures.ts';
